w3c-reports/12092014/mac/runned-tests/html/semantics/embedded-content/media-elements/track/track-element/cloneNode-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.

PASS track element cloneNode, not loaded 
PASS track element cloneNode, loading 
FAIL track element cloneNode, loaded assert_unreached: elm got error Reached unreachable code(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_unreached@http://localhost:8800/resources/testharness.js:692:16
http://localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cloneNode.html:66:63
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS track element cloneNode, failed to load 
